maplibre-gl-js updated to 4.5.0 meshpipe.py update to support backend trackMessage and sqlite storage of other node information gpsreader update for manual positioning delivery to UI fifo pipe
Manual location first prototype now implemented. Allows user to select manual location when locally
connected GPS is unusable. Removed browser location, periodic send from UI and removed peers button.

+++ board/raspberrypi/post-build.sh
@@ -11,7 +11,14 @@ tty1::respawn:/sbin/getty -L  tty1 0 vt100 # HDMI console' ${TARGET_DIR}/etc/ini
 fi
 
 # Remove janus-gateway.service
+if [ -f ${TARGET_DIR}/usr/lib/systemd/system/janus-gateway.service ]; then
 rm ${TARGET_DIR}/usr/lib/systemd/system/janus-gateway.service
+fi
 
 # Alter permissions to /opt/edgemap-persist
+if [ -d ${TARGET_DIR}/opt/edgemap-persist ]; then
 chmod o+w ${TARGET_DIR}/opt/edgemap-persist
+else
+mkdir ${TARGET_DIR}/opt/edgemap-persist
+chmod o+w ${TARGET_DIR}/opt/edgemap-persist
+fi
